% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxMinPollutedCity.R
\name{maxMinPollutedCity}
\alias{maxMinPollutedCity}
\title{TOP AQIs: WHERE and HOW HIGH}
\usage{
maxMinPollutedCity(num)
}
\arguments{
\item{num}{}
}
\value{
The current top "num" worst AQIs in the US, where "num" is an integer from 1 to 5
}
\description{
TOP AQIs: WHERE and HOW HIGH
}
\examples{
maxMinPolutedCity(1)
maxMinPolutedCity(3)
}
