% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_air.R
\name{zipAir}
\alias{zipAir}
\title{most recent Air Quality in a zip code.}
\usage{
zipAir(zip)
}
\arguments{
\item{zip}{}
}
\value{
The current Air Quality Index of a certain zip code along with any associated health messages.
}
\description{
most recent Air Quality in a zip code.
}
\examples{
zipAir(90007)
zipAir(80021)
}
